Adding final implementation of movieDetail screen except review screen

+++ app/src/main/java/com/example/assignment_for_daresay/data/model/Reviews.kt
@@ -27,5 +27,5 @@ data class AuthorDetails (
     val name: String?,
     val username : String?,
     val avatar_path : String? = null,
-    val rating : Int? = null
+    val rating : Double? = null
 )
